backend/pyproject.toml
[project]
name = "echoforge-backend"
version = "0.1.0"
description = "EchoForge FastAPI backend scaffold"
requires-python = ">=3.11"
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "sqlalchemy>=2.0",
    "psycopg[binary]>=3.1",
    "pyyaml>=6.0",
    "alembic>=1.13",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3",
]

[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]
exclude = ["tests*"]

[tool.poetry]
package-mode = false
